ENH: Refactor xarray support into separate module
